Add devel::rustc v1.81.0
This includes a target for the rustc + std lib + tools. It is also
possible to install additional std libs for other architectures. Finally
there are targets to install a rustc + extra std libs which can be used
as tools while building. x86_64, i386, aarch64 and armv7 are supported
right now.